server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
requests==2.31.0
pymongo==4.7.2
dnspython==2.5.0
python-dotenv==1.0.1
